+++ react/components/modals/UserConfirmationModal.tsx
@@ -35,7 +35,9 @@ const UserConfirmationModal = ({
       }}
       cancelation={{
         onClick: () => onClose(),
-        label: intl.formatMessage({ id: 'store/my-users.button.cancel' }),
+        label: intl.formatMessage({
+          id: 'store/my-organization.button.cancel',
+        }),
       }}
       isOpen={isOpen}
       onClose={() => onClose()}>
@@ -43,9 +45,7 @@ const UserConfirmationModal = ({
       <div className="flex flex-column mb5 mt5">
         <div>
           <span>{message}</span>
-          <span className="b">
-            {pathOr('', ['personaId_linked', 'email'], assignment)}
-          </span>
+          <span className="b">{pathOr('', ['email'], assignment)}</span>
         </div>
       </div>
     </ModalDialog>