+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- Message added when user is not assigned to any organization
+
 ## [1.2.0] - 2020-10-29
 
 ## [1.1.2] - 2020-10-29