+++ CHANGELOG.md
@@ -6,7 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- Change labels, use `My Organization` / `organization` instead of `My users`/`users`
+### Changed
+- Update labels to use `My Organization` / `organization` instead of `My users`/`users`
 
 ## [1.6.0] - 2020-12-10
 
@@ -81,4 +82,3 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
     * `Persona` with schema name `persona-schema-v1`
     * `BusinessOrganization` with schema name `business-organization-schema-v1`
     * `OrgAssignment` with schema name `organization-assignment-schema-v1`
-